refactor: staging is sematically different from storage
